Commit changes; I believe proving backwards simulation is pointless
So the new proof idea was the following:

Rule out the cases
1) p**
2) all*
3) []

And move on the the following: ~ (p* match s) will give
translate(p*, r) match s false because if you "run out of characters",
p* fails after consuming all characters, and so
1) translate(p*, emp) will succeed with [] and no other characters and
2) r will fail on this consumed all characters string

2) we probably have to add as a hypothesis; an interesting way to do
it might be making r a terminating character; then we would simply
have to prove that it holds inductively, that is, if r does not match
consumes_all_chars(p), then forall p1, translate(p, r) will not match
consumesallchars(p1).
The above inductively, does not hold. But maybe it will hold if, for
example, you rule out expressions like
p*p, which are also unmatchable. In fact, we ought to rule out any peg
where p1.p2 & p2 matches all of the characters of a failing read of
p1. Then maybe it will hold inductively.